# Reduce memory footprint even lower